Choosing us component as it's own component
