+++ src/component/PropertyCardComponent.js
@@ -39,13 +39,6 @@ const CategoryCardComponent = ({ category, id }) => {
         },
     ];
 
-    const whyChooseUs = [
-        { icon: 'bi bi-send', title: 'Service', description: 'We deliver high quality and exceptional service to our clients.' },
-        { icon: 'bi bi-eject', title: 'Experience', description: 'Over 10 years in providing professional real estate solutions.' },
-        { icon: 'bi bi-house-gear', title: 'Our Staff', description: 'Highly motivated and dedicated to meet your real estate needs.' },
-
-    ]
-
     return (
         <Container>
             <h4 style={{ marginTop: "10px", marginBottom: "10px", textAlign: "center" }}>Find Recent Properties</h4>
@@ -69,23 +62,6 @@ const CategoryCardComponent = ({ category, id }) => {
                         </Col>
                     );
                 })}
-            </Row>
-            <h4 style={{ marginTop: "10px", marginBottom: "10px", textAlign: "center" }}>Why choosing us</h4>
-            <Row>
-                {whyChooseUs.map((item) => {
-                    return (
-                        <Col md={4} sm={6}>
-                            <Card className="text-center">
-                                <Card.Body>
-                                    <i className={item.icon} style={{ fontSize: "30px" }}></i>
-                                    <Card.Title style={{ fontSize: "30px" }}>{item.title}</Card.Title>
-                                    <Card.Text>{item.description}</Card.Text>
-                                </Card.Body>
-                            </Card>
-                        </Col>
-                    )
-                })}
-
             </Row>
         </Container >
     );
