+++ src/pages/Home.jsx
@@ -2,6 +2,7 @@ import React from "react";
 import FooterComponent from "../component/FooterComponent";
 
 import PropertyCarouselComponent from "../component/PropertyCarouselComponent";
+import ChoosingUsComponent from "../component/ChoosingUsComponent";
 import HouseCardComponent from "../component/PropertyCardComponent";
 
 const Home = () => {
@@ -9,6 +10,7 @@ const Home = () => {
     <div>
       <PropertyCarouselComponent />
       <HouseCardComponent />
+      <ChoosingUsComponent />
       <FooterComponent />
     </div>
   );
